@@ -1,11 +1,13 @@
 import React, { useContext, useEffect } from 'react';
+import { useSelector } from 'react-redux';
 
 import Layout from '../../layout/';
 
 import { CommonContext } from '../../context/CommonContext';
 
 const Auth = () => {
-  const { user, setSignDialogOpen, setUserDetailDialogOpen } = useContext(
+  const user = useSelector(state => state.Auth.user);
+  const { setSignDialogOpen, setUserDetailDialogOpen } = useContext(
     CommonContext,
   );
 
